+++ kyuubi-server/src/main/scala/org/apache/kyuubi/server/BackendServiceMetric.scala
@@ -25,7 +25,7 @@ import org.apache.kyuubi.operation.FetchOrientation.FetchOrientation
 import org.apache.kyuubi.service.BackendService
 import org.apache.kyuubi.session.SessionHandle
 
-trait BackendServiceTimeMetric extends BackendService {
+trait BackendServiceMetric extends BackendService {
 
   abstract override def openSession(
       protocol: TProtocolVersion,
@@ -152,7 +152,28 @@ trait BackendServiceTimeMetric extends BackendService {
       maxRows: Int,
       fetchLog: Boolean): TRowSet = {
     MetricsSystem.timerTracing(MetricsConstants.BS_FETCH_RESULTS) {
-      super.fetchResults(operationHandle, orientation, maxRows, fetchLog)
+      val rowSet = super.fetchResults(operationHandle, orientation, maxRows, fetchLog)
+      val rowsSize =
+        if (rowSet.getColumnsSize > 0) {
+          rowSet.getColumns.get(0).getFieldValue match {
+            case t: TStringColumn => t.getValues.size()
+            case t: TDoubleColumn => t.getValues.size()
+            case t: TI64Column => t.getValues.size()
+            case t: TI32Column => t.getValues.size()
+            case t: TI16Column => t.getValues.size()
+            case t: TBoolColumn => t.getValues.size()
+            case t: TByteColumn => t.getValues.size()
+            case t: TBinaryColumn => t.getValues.size()
+            case _ => 0
+          }
+        } else rowSet.getRowsSize
+
+      MetricsSystem.tracing(_.markMeter(
+        if (fetchLog) MetricsConstants.BS_FETCH_LOG_ROWS_RATE
+        else MetricsConstants.BS_FETCH_RESULT_ROWS_RATE,
+        rowsSize))
+
+      rowSet
     }
   }
