+++ kyuubi-server/src/test/scala/org/apache/kyuubi/server/BackendServiceMetricSuite.scala
@@ -27,7 +27,7 @@ import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.metrics.{MetricsConf, MetricsConstants}
 import org.apache.kyuubi.operation.HiveJDBCTestHelper
 
-class BackendServiceTimeMetricSuite extends WithKyuubiServer with HiveJDBCTestHelper {
+class BackendServiceMetricSuite extends WithKyuubiServer with HiveJDBCTestHelper {
 
   override protected def jdbcUrl: String = getJdbcUrl
 
@@ -39,35 +39,49 @@ class BackendServiceTimeMetricSuite extends WithKyuubiServer with HiveJDBCTestHe
       .set(MetricsConf.METRICS_JSON_INTERVAL, Duration.ofMillis(100).toMillis)
   }
 
-  test("backend service method time metric test") {
+  test("backend service metric test") {
     val objMapper = new ObjectMapper()
 
     withJdbcStatement() { statement =>
-      statement.execute("show databases")
+      statement.executeQuery("CREATE TABLE stu_test(id int, name string) USING parquet")
+      statement.execute("insert into stu_test values(1, 'a'), (2, 'b'), (3, 'c')")
       Thread.sleep(Duration.ofMillis(111).toMillis)
 
       val res1 = objMapper.readTree(Paths.get(reportPath.toString, "report.json").toFile)
       assert(res1.has("timers"))
-      val histograms1 = res1.get("timers")
+      val timer1 = res1.get("timers")
       assert(
-        histograms1.get(MetricsConstants.BS_EXECUTE_STATEMENT).get("count").asInt() == 1)
+        timer1.get(MetricsConstants.BS_EXECUTE_STATEMENT).get("count").asInt() == 2)
       assert(
-        histograms1.get(MetricsConstants.BS_EXECUTE_STATEMENT).get("mean").asDouble() > 0)
+        timer1.get(MetricsConstants.BS_EXECUTE_STATEMENT).get("mean").asDouble() > 0)
 
-      statement.execute("show tables")
+      assert(res1.has("meters"))
+      val meters1 = res1.get("meters")
+      val logRows1 = meters1.get(MetricsConstants.BS_FETCH_LOG_ROWS_RATE).get("count").asInt()
+      assert(logRows1 > 0)
+
+      statement.execute("select * from stu_test limit 2")
+      statement.getResultSet.next()
       Thread.sleep(Duration.ofMillis(111).toMillis)
 
       val res2 = objMapper.readTree(Paths.get(reportPath.toString, "report.json").toFile)
-      val histograms2 = res2.get("timers")
+      val timer2 = res2.get("timers")
       assert(
-        histograms2.get(MetricsConstants.BS_OPEN_SESSION).get("count").asInt() == 1)
+        timer2.get(MetricsConstants.BS_OPEN_SESSION).get("count").asInt() == 1)
       assert(
-        histograms2.get(MetricsConstants.BS_OPEN_SESSION).get("min").asInt() > 0)
-      val execStatementNode2 = histograms2.get(MetricsConstants.BS_EXECUTE_STATEMENT)
-      assert(execStatementNode2.get("count").asInt() == 2)
+        timer2.get(MetricsConstants.BS_OPEN_SESSION).get("min").asInt() > 0)
+      val execStatementNode2 = timer2.get(MetricsConstants.BS_EXECUTE_STATEMENT)
+      assert(execStatementNode2.get("count").asInt() == 3)
       assert(
         execStatementNode2.get("max").asDouble() >= execStatementNode2.get("mean").asDouble() &&
           execStatementNode2.get("mean").asDouble() >= execStatementNode2.get("min").asDouble())
+
+      val meters2 =
+        objMapper.readTree(Paths.get(reportPath.toString, "report.json").toFile).get("meters")
+      assert(meters2.get(MetricsConstants.BS_FETCH_RESULT_ROWS_RATE).get("count").asInt() == 2)
+      assert(meters2.get(MetricsConstants.BS_FETCH_LOG_ROWS_RATE).get("count").asInt() >= logRows1)
+
+      statement.executeQuery("DROP TABLE stu_test")
     }
   }
 }
